Refactored page layout
I refactored the the application to allow for easy construction of
different pages. In addition, I added a general use menuListener for all
of the menu buttons we will add for swapping to differnt pages. will
likely add the other pages first before trying to delve into connecting
to database.