Overhauled the Looks
I wrote a renderer for making the listboxes display in a different color
and for me to be able to specify the selection color to them. I also
changed the entire look and feel of the program to be a lot darker
because it looks better imo. all of the colors are linked to constants
in main so if we want to change the feel later, only need to change 5
lines of code. Also made login listener extend keyListener so I could
add it to the Password field to make the enter button work.
Additionally, I finished making the Restock page and decided to go with
a set 20 item boxes per restock because the box layout manager looked
terrible and the only other layout manager that would work well was
grid, which requires that you specify a set number of rows at creation.
next will probably work on the game text parsing system. (though not
sure yet when)