fixed an issue with double refresh on select
fixed and issue with the data being refreshed twice when selecting a
shop in the shopList. had to do with an event being sent to the listener
twice for every select, one for an intermeadiate state, and one for the
finished state. simpily only have the refresh happen for the finished
state event.